+++ jdaviz/configs/default/plugins/gaussian_smooth/gaussian_smooth.vue
@@ -37,6 +37,21 @@
           ></v-text-field>
         </v-col>
       </v-row>
+      <v-row v-if="selected_data">
+        <v-select v-if="config=='cubeviz'"
+          :items="viewers"
+          v-model="selected_viewer"
+          label='Plot in Viewer'
+          hint='Smoothed data will replace plot in the specified viewer.  Will also be available in the data dropdown in all image viewers.'
+          persistent-hint></v-select>
+        <v-switch v-if="selected_mode !== 'Spatial'"
+          label="Plot in Viewer"
+          hint="Smoothed data will immediately plot in the spectral viewer.  Will also be available in the data menu of each viewer."
+          v-model="add_replace_results"
+          persistent-hint>
+        </v-switch>
+        </div>
+      </v-row>
     </v-container>
     <!-- <v-divider></v-divider> -->